refactor: extract input-container logic into reusable mixin
